requirements.txt
# MemISTD Requirements for Linux Server
# Python >= 3.11

# ===========================================
# PyTorch (CUDA 11.8 version for Linux GPU servers)
# Note: For CPU-only servers, use torch without CUDA index
# ===========================================
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.2
torchvision==0.16.2

# ===========================================
# Core Deep Learning Dependencies
# ===========================================
numpy>=1.24.0,<2.0.0
scipy>=1.11.0

# ===========================================
# Image Processing
# ===========================================
Pillow>=10.0.0
opencv-python>=4.8.0

# ===========================================
# Configuration and Utilities
# ===========================================
PyYAML>=6.0
tqdm>=4.65.0
einops>=0.7.0

# ===========================================
# Visualization and Logging
# ===========================================
matplotlib>=3.7.0
tensorboard>=2.14.0

# ===========================================
# COCO Evaluation Tools
# ===========================================
pycocotools>=2.0.7
